fix(CartPresenter): 포지션 값이 양수일 때만 아이템을 제거하도록 수정

+++ app/src/main/java/woowacourse/shopping/presentation/view/cart/CartPresenter.kt
@@ -13,8 +13,10 @@ class CartPresenter(
     }
 
     override fun deleteCartItem(position: Int) {
-        cartRepository.deleteCartByProductId(carts[position].product.id)
-        carts.removeAt(position)
-        view.updateToDeleteCartItemView(position)
+        if (position >= 0) {
+            cartRepository.deleteCartByProductId(carts[position].product.id)
+            carts.removeAt(position)
+            view.updateToDeleteCartItemView(position)
+        }
     }
 }
